requirements.txt
matplotlib~=3.8.3
torch~=2.2.1
torchvision~=0.17.1
numpy~=1.26.4
pillow~=10.2.0
tqdm~=4.66.2
scikit-learn~=1.4.1.post1